chore: move refund_parameter types to own file.
