+++ api/cf-node-client/src/subxt_state_chain_config.rs
@@ -68,7 +68,7 @@ include!(concat!(env!("OUT_DIR"), "/cf_static_runtime.rs"));
 // Conversions from cf_static_runtime::runtime_types
 impl<T>
 	From<
-		cf_static_runtime::runtime_types::cf_chains::ChannelRefundParametersGeneric<
+		cf_static_runtime::runtime_types::cf_chains::refund_parameters::ChannelRefundParametersGeneric<
 			T,
 			Option<
 				cf_static_runtime::runtime_types::cf_chains::CcmChannelMetadata<
@@ -79,7 +79,7 @@ impl<T>
 	> for cf_chains::ChannelRefundParametersGeneric<T>
 {
 	fn from(
-		value: cf_static_runtime::runtime_types::cf_chains::ChannelRefundParametersGeneric<
+		value: cf_static_runtime::runtime_types::cf_chains::refund_parameters::ChannelRefundParametersGeneric<
 			T,
 			Option<
 				cf_static_runtime::runtime_types::cf_chains::CcmChannelMetadata<
